Staging

+++ app/Seggu.Dtos/AccessoryDto.cs
@@ -8,7 +8,7 @@ namespace Seggu.Dtos
         public string Oblea { get; set; }
         public DateTime Vencimiento { get; set; }
         public decimal Valor { get; set; }
-        public int VehicleId { get; set; }
-        public int AccessoryTypeId { get; set; }
+        public long VehicleId { get; set; }
+        public long AccessoryTypeId { get; set; }
     }
 }
